store/chrome: link to NSS for cert store on linux
Issue: https://github.com/adamdecaf/cert-manage/issues/10
