+++ store/chrome_linux.go
@@ -2,8 +2,16 @@
 
 package store
 
+// TODO(adam): impl
+func chromeCert8Locations() []string {
+	return nil
+}
+
 // On linux chrome uses NSS
 // https://www.chromium.org/Home/chromium-security/root-ca-policy
 func ChromeStore() Store {
-	return NssStore()
+	suggestions := collectNssSuggestions(chromeCert8Locations())
+	return nssStore{
+		paths: suggestions,
+	}
 }
